backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend"
requires-python = ">=3.12"

dependencies = [
    "aegra-cli>=0.8.2",
    "langgraph>=1.0.8",
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.12",
    "python-multipart>=0.0.22",
    "llama-index-readers-file>=0.6.0",
    "llama-index-vector-stores-postgres>=0.8.1",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-openai>=0.7.3",
    "langchain-community>=0.4.1",
    "tavily-python>=0.7.23",
    "llama-cloud>=2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]
